@@ -8,42 +8,42 @@
         "node_modules"
     ],
     "dependencies": {
-        "@angular/animations": "14.1.2",
-        "@angular/cdk": "14.1.2",
-        "@angular/common": "14.1.2",
-        "@angular/compiler": "14.1.2",
-        "@angular/core": "14.1.2",
-        "@angular/forms": "14.1.2",
-        "@angular/localize": "14.1.2",
-        "@angular/material": "14.1.2",
-        "@angular/platform-browser": "14.1.2",
-        "@angular/platform-browser-dynamic": "14.1.2",
-        "@angular/router": "14.1.2",
-        "@angular/service-worker": "14.1.2",
+        "@angular/animations": "14.2.0",
+        "@angular/cdk": "14.2.0",
+        "@angular/common": "14.2.0",
+        "@angular/compiler": "14.2.0",
+        "@angular/core": "14.2.0",
+        "@angular/forms": "14.2.0",
+        "@angular/localize": "14.2.0",
+        "@angular/material": "14.2.0",
+        "@angular/platform-browser": "14.2.0",
+        "@angular/platform-browser-dynamic": "14.2.0",
+        "@angular/router": "14.2.0",
+        "@angular/service-worker": "14.2.0",
         "@ctrl/ngx-emoji-mart": "7.0.0",
         "@danielmoncada/angular-datetime-picker": "14.1.1",
-        "@fingerprintjs/fingerprintjs": "3.3.4",
+        "@fingerprintjs/fingerprintjs": "3.3.5",
         "@flaviosantoro92/ngx-datatable": "1.0.2",
         "@fortawesome/angular-fontawesome": "0.11.1",
-        "@fortawesome/fontawesome-svg-core": "6.1.2",
-        "@fortawesome/free-regular-svg-icons": "6.1.2",
-        "@fortawesome/free-solid-svg-icons": "6.1.2",
-        "@ls1intum/apollon": "2.12.1",
+        "@fortawesome/fontawesome-svg-core": "6.2.0",
+        "@fortawesome/free-regular-svg-icons": "6.2.0",
+        "@fortawesome/free-solid-svg-icons": "6.2.0",
+        "@ls1intum/apollon": "2.12.2",
         "@ng-bootstrap/ng-bootstrap": "13.0.0",
         "@ngx-translate/core": "14.0.0",
         "@ngx-translate/http-loader": "7.0.0",
-        "@sentry/angular": "7.10.0",
-        "@sentry/tracing": "7.10.0",
+        "@sentry/angular": "7.12.1",
+        "@sentry/tracing": "7.12.1",
         "@swimlane/ngx-charts": "20.1.0",
-        "ace-builds": "1.9.5",
+        "ace-builds": "1.10.0",
         "bootstrap": "5.2.0",
         "brace": "0.11.1",
-        "compare-versions": "4.1.3",
-        "core-js": "3.24.1",
+        "compare-versions": "5.0.1",
+        "core-js": "3.25.0",
         "crypto-js": "4.1.1",
         "dayjs": "1.11.5",
         "diff-match-patch-typescript": "1.0.8",
-        "dompurify": "2.3.10",
+        "dompurify": "2.4.0",
         "export-to-csv": "0.2.1",
         "franc-min": "6.0.0",
         "interactjs": "1.10.17",
@@ -76,34 +76,34 @@
         }
     },
     "devDependencies": {
-        "@angular-builders/custom-webpack": "14.0.0",
-        "@angular-builders/jest": "14.0.0",
-        "@angular-devkit/build-angular": "14.1.2",
-        "@angular-eslint/eslint-plugin": "14.0.2",
-        "@angular/cli": "14.1.2",
-        "@angular/compiler-cli": "14.1.2",
-        "@angular/language-service": "14.1.2",
+        "@angular-builders/custom-webpack": "14.0.1",
+        "@angular-builders/jest": "14.0.1",
+        "@angular-devkit/build-angular": "14.2.1",
+        "@angular-eslint/eslint-plugin": "14.0.3",
+        "@angular/cli": "14.2.1",
+        "@angular/compiler-cli": "14.2.0",
+        "@angular/language-service": "14.2.0",
         "@types/crypto-js": "4.1.1",
-        "@types/dompurify": "2.3.3",
-        "@types/jest": "28.1.6",
+        "@types/dompurify": "2.3.4",
+        "@types/jest": "28.1.8",
         "@types/lodash-es": "4.17.6",
-        "@types/node": "18.7.2",
-        "@types/papaparse": "5.3.3",
+        "@types/node": "18.7.15",
+        "@types/papaparse": "5.3.5",
         "@types/showdown": "2.0.0",
         "@types/smoothscroll-polyfill": "0.3.1",
         "@types/sockjs-client": "1.5.1",
         "@types/uuid": "8.3.4",
-        "@typescript-eslint/eslint-plugin": "5.33.0",
-        "@typescript-eslint/eslint-plugin-tslint": "5.33.0",
-        "@typescript-eslint/parser": "5.33.0",
+        "@typescript-eslint/eslint-plugin": "5.36.1",
+        "@typescript-eslint/eslint-plugin-tslint": "5.36.1",
+        "@typescript-eslint/parser": "5.36.1",
         "browser-sync": "2.27.10",
         "browser-sync-webpack-plugin": "2.3.0",
         "codelyzer": "6.0.2",
-        "eslint": "8.21.0",
+        "eslint": "8.23.0",
         "eslint-config-prettier": "8.5.0",
         "eslint-plugin-deprecation": "1.3.2",
-        "eslint-plugin-jest": "26.8.2",
-        "eslint-plugin-jest-extended": "1.0.0",
+        "eslint-plugin-jest": "27.0.1",
+        "eslint-plugin-jest-extended": "2.0.0",
         "eslint-plugin-prettier": "4.2.1",
         "eslint-webpack-plugin": "3.2.0",
         "folder-hash": "4.0.2",
@@ -111,22 +111,22 @@
         "jest": "28.1.3",
         "jest-canvas-mock": "2.4.0",
         "jest-date-mock": "1.0.8",
-        "jest-extended": "3.0.2",
-        "jest-junit": "14.0.0",
-        "jest-preset-angular": "12.2.0",
+        "jest-extended": "3.1.0",
+        "jest-junit": "14.0.1",
+        "jest-preset-angular": "12.2.2",
         "lint-staged": "13.0.3",
         "merge-jsons-webpack-plugin": "2.0.1",
-        "ng-mocks": "14.1.3",
+        "ng-mocks": "14.2.0",
         "prettier": "2.7.1",
-        "sass": "1.54.4",
-        "ts-jest": "28.0.7",
+        "sass": "1.54.8",
+        "ts-jest": "28.0.8",
         "tslint": "6.1.3",
         "tslint-config-prettier": "1.18.0",
-        "typescript": "4.7.4",
+        "typescript": "4.8.2",
         "weak-napi": "2.0.2",
         "webpack": "5.74.0",
-        "webpack-bundle-analyzer": "4.5.0",
-        "webpack-dev-server": "4.10.0",
+        "webpack-bundle-analyzer": "4.6.1",
+        "webpack-dev-server": "4.10.1",
         "webpack-merge": "5.8.0",
         "webpack-notifier": "1.15.0"
     },
