@@ -371,7 +371,7 @@ export class FeatureOverviewComponent implements OnInit {
             'featureOverview.instructor.feature.liveStatistics.descriptionTextOne',
             faEye,
             undefined,
-            '/content/images/feature-overview/instructors/exam_live_statistics.png',
+            '/content/images/feature-overview/instructors/exam_live_statistics_section_overview_section.png',
         );
 
         this.features = [
