@@ -1,11 +1,30 @@
-// NOTE: this list has to be kept in sync with the extensions defined in FileResource.java
+/**
+ * The list of file extensions that are allowed to be uploaded in a Markdown editor.
+ * Extensions must be lower-case without leading dots.
+ * NOTE: Has to be kept in sync with the server-side definitions in FileResource.java
+ */
+export const MARKDOWN_FILE_EXTENSIONS = ['png', 'jpg', 'jpeg', 'gif', 'svg', 'pdf'];
+
+/**
+ * The global list of file extensions that are allowed to be uploaded.
+ * Extensions must be lower-case without leading dots.
+ * NOTE: Has to be kept in sync with the server-side definitions in FileResource.java
+ */
 export const FILE_EXTENSIONS = [
     'png',
     'jpg',
     'jpeg',
+    'gif',
     'svg',
     'pdf',
     'zip',
+    'tar',
+    'txt',
+    'rtf',
+    'md',
+    'htm',
+    'html',
+    'json',
     'doc',
     'docx',
     'csv',
@@ -14,7 +33,6 @@ export const FILE_EXTENSIONS = [
     'ppt',
     'pptx',
     'pages',
-    'rtf',
     'pages-tef',
     'numbers',
     'key',
@@ -22,5 +40,7 @@ export const FILE_EXTENSIONS = [
     'ods',
     'odp',
     'odg',
+    'odc',
+    'odi',
     'odf',
 ];
