@@ -1,5 +1,5 @@
 import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { ActivatedRoute, Router, RouterOutlet } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { NgbPagination } from '@ng-bootstrap/ng-bootstrap';
 import { SystemNotificationManagementComponent } from 'app/admin/system-notification-management/system-notification-management.component';
 import { SystemNotification } from 'app/entities/system-notification.model';
@@ -36,7 +36,6 @@ describe('SystemNotificationManagementComponent', () => {
                 MockPipe(ArtemisDatePipe),
                 MockPipe(ArtemisTranslatePipe),
                 MockRouterLinkDirective,
-                MockDirective(RouterOutlet),
                 MockDirective(DeleteButtonDirective),
                 MockDirective(SortDirective),
                 MockComponent(ItemCountComponent),
